refactor: transform science module
