+++ src/modules/science/index.ts
@@ -1,1 +1,3 @@
+export type { ChemicalElement } from './chemical-element';
 export * from './module';
+export type { Unit } from './unit';
