+++ src/modules/science/module.ts
@@ -1,33 +1,6 @@
 import { ModuleBase } from '../../internal/module-base';
-
-/**
- * The possible definitions related to elements.
- */
-export interface ChemicalElement {
-  /**
-   * The symbol for the element (e.g. `'He'`).
-   */
-  symbol: string;
-  /**
-   * The name for the element (e.g. `'Cerium'`).
-   */
-  name: string;
-  /**
-   * The atomic number for the element (e.g. `52`).
-   */
-  atomicNumber: number;
-}
-
-export interface Unit {
-  /**
-   * The long version of the unit (e.g. `meter`).
-   */
-  name: string;
-  /**
-   * The short version/abbreviation of the unit (e.g. `Pa`).
-   */
-  symbol: string;
-}
+import type { ChemicalElement } from './chemical-element';
+import type { Unit } from './unit';
 
 /**
  * Module to generate science related entries.
